pyproject.toml
[project]
name = "calliope-cli"
version = "0.1.0"
description = "Calliope CLI: a writing/reading (拆书) agent with chat UI and RAG."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "typer==0.20.0",
    "kosong==0.25.0",
    "loguru==0.7.3",
    "pydantic==2.12.4",
    "rich==14.2.0",
    "prompt-toolkit==3.0.52",
    "aiofiles==25.1.0",
    "pyyaml==6.0.3",
    "httpx[socks]==0.28.1",
]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.4",
]

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "calliope_cli"
source-exclude = ["tests/**/*"]

[project.scripts]
calliope = "calliope_cli.cli:cli"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "F",
    "UP",
    "B",
    "SIM",
    "I",
]
